requirements.txt
gradio>=4.0.0
pillow
numpy